rename TestSign.vue to Sign.vue and delete SignIn SignUp components
